% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ties.workshops-package.R
\docType{package}
\name{ties.workshops-package}
\alias{ties.workshops}
\alias{ties.workshops-package}
\title{ties.workshops: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
